+++ lib/controller/order_controller.dart
@@ -1,11 +1,12 @@
 part of '_controller.dart';
 
 class OrderController extends GetxController {
+  bool isOrdered = false;
   OrderResponse? _orderResponse;
 
   OrderResponse? get getOrderResponse => _orderResponse;
 
-  set setOrderResponse(OrderResponse value) {
+  set setOrderResponse(OrderResponse? value) {
     _orderResponse = value;
     update();
   }
