+++ lib/helpers/order_helper.dart
@@ -39,7 +39,7 @@ class OrderHelper {
     );
     print('ORDER SUCCESSFULLY CREATED');
 
-    // final cart = Provider.of<CartProvider>(_context, listen: false);
-    // cart.items=[];
+    final cart = Provider.of<CartProvider>(_context, listen: false);
+    cart.items=[];
   }
 }
